requirements.txt
streamlit
streamlit-webrtc
opencv-python
mediapipe
Pillow
